+++ src/features/book-note/ui/BookNoteForm/QuotesStep/index.tsx
