+++ src/features/book-note/ui/BookNoteForm/ReadingInfoStep/BookInfo.tsx
