--- src/features/book-note/ui/BookNoteForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BookNoteForm';
